chore: export factory from typography
